Initial support for webpack@2.
This is obviously a breaking change because of the changes to the webpack API. Smoke tests have been added to ensure it works in the most basic of cases. `postcss-import` is pinned to `8.1.0` because of a breaking bug with `webpack@2`. See: https://github.com/postcss/postcss-import/issues/207.

Support for `node@0.12` is removed because `webpack@2` does not support it.